chore: rename lib/common/stdexcept.h => lib/common/StdExcept.h
