+++ src/lib/deskflow/StreamChunker.cpp
@@ -12,7 +12,7 @@
 #include "base/Log.h"
 #include "base/Stopwatch.h"
 #include "base/String.h"
-#include "common/stdexcept.h"
+#include "common/StdExcept.h"
 #include "deskflow/ClipboardChunk.h"
 #include "deskflow/FileChunk.h"
 #include "deskflow/ProtocolTypes.h"
