+++ src/lib/server/Server.cpp
@@ -12,7 +12,7 @@
 #include "base/Log.h"
 #include "base/TMethodEventJob.h"
 #include "base/TMethodJob.h"
-#include "common/stdexcept.h"
+#include "common/StdExcept.h"
 #include "deskflow/AppUtil.h"
 #include "deskflow/DropHelper.h"
 #include "deskflow/FileChunk.h"
