+++ src/test/integtests/net/NetworkTests.cpp
@@ -14,7 +14,7 @@
 #include "base/TMethodEventJob.h"
 #include "base/TMethodJob.h"
 #include "client/Client.h"
-#include "common/stdexcept.h"
+#include "common/StdExcept.h"
 #include "deskflow/FileChunk.h"
 #include "deskflow/StreamChunker.h"
 #include "mt/Thread.h"
